+++ src/elements/conditions/BlockCondition.php
@@ -2,8 +2,10 @@
 
 namespace benf\neo\elements\conditions;
 
+use benf\neo\Plugin as Neo;
 use craft\elements\conditions\ElementCondition;
 use craft\elements\conditions\LevelConditionRule;
+use craft\models\FieldLayout;
 
 /**
  * Class BlockCondition
@@ -22,22 +24,37 @@ class BlockCondition extends ElementCondition
         $parentConditionRuleTypes = parent::selectableConditionRules();
         $fieldConditionRuleTypes = [];
 
-        foreach ($parentConditionRuleTypes as $ruleType) {
-            if (!isset($ruleType['class'])) {
-                continue;
-            }
+        // Get all field layouts associated with this object's associated Neo field(s), then temporarily replace this
+        // object's field layouts so we get all possible parent block condition rules
+        $layoutBlockTypes = Neo::$plugin->blockTypes->getByCriteria([
+            'fieldLayoutId' => array_map(fn($layout) => $layout->id, $this->getFieldLayouts()),
+        ]);
+        $fieldBlockTypes = Neo::$plugin->blockTypes->getByCriteria([
+            'fieldId' => array_values(array_unique(array_map(fn($blockType) => $blockType->fieldId, $layoutBlockTypes))),
+        ]);
+        $fieldLayouts = array_map(fn($blockType) => $blockType->getFieldLayout(), $fieldBlockTypes);
+        $fieldConditionRuleTypes = array_values(array_filter(array_map(
+            function($ruleType) {
+                if (!isset($ruleType['class'])) {
+                    return null;
+                }
 
-            $splitClass = explode('\\', $ruleType['class']);
-            $className = __NAMESPACE__ . '\\fields\\Parent' . end($splitClass);
+                $splitClass = explode('\\', $ruleType['class']);
+                $className = __NAMESPACE__ . '\\fields\\Parent' . end($splitClass);
 
-            if (class_exists($className)) {
-                $fieldConditionRuleTypes[] = [
-                    'class' => $className,
-                    'fieldUid' => $ruleType['fieldUid'],
-                    'layoutElementUid' => $ruleType['layoutElementUid'],
-                ];
-            }
-        }
+                if (class_exists($className)) {
+                    return [
+                        'class' => $className,
+                        'fieldUid' => $ruleType['fieldUid'],
+                        'layoutElementUid' => $ruleType['layoutElementUid'],
+                    ];
+                }
+            },
+            $this->_swapFieldLayoutsWithThen(
+                $fieldLayouts,
+                fn() => parent::selectableConditionRules(),
+            ),
+        )));
 
         return array_merge(
             $parentConditionRuleTypes,
@@ -52,5 +69,26 @@ class BlockCondition extends ElementCondition
                 OwnerVolumeConditionRule::class,
             ],
         );
+    }
+
+    /**
+     * Temporarily swaps the condition field layouts before calling a given function.
+     *
+     * @param FieldLayout[]|null $with An array of field layouts to swap with the condition field layouts
+     * @param callable $then A function to run while the condition field layouts are swapped
+     * @return mixed The return value from $then
+     */
+    private function _swapFieldLayoutsWithThen(?array $with, callable $then): mixed
+    {
+        if ($with) {
+            $fieldLayouts = $this->getFieldLayouts();
+            $this->setFieldLayouts($with);
+            $returnVal = $then();
+            $this->setFieldLayouts($fieldLayouts);
+
+            return $returnVal;
+        }
+
+        return $then();
     }
 }